+++ android/src/main/java/com/revenuecat/purchases_flutter/PurchasesFlutterPlugin.java
@@ -39,6 +39,8 @@ import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
 import io.flutter.plugin.common.MethodChannel.Result;
 import kotlin.UninitializedPropertyAccessException;
+import kotlin.Unit;
+import kotlin.jvm.functions.Function1;
 
 /**
  * PurchasesFlutterPlugin
@@ -60,7 +62,7 @@ public class PurchasesFlutterPlugin implements FlutterPlugin, MethodCallHandler,
     private final Handler handler = new Handler(Looper.getMainLooper());
 
     private static final String PLATFORM_NAME = "flutter";
-    private static final String PLUGIN_VERSION = "8.7.4";
+    private static final String PLUGIN_VERSION = "8.10.0";
 
     @Override
     public void onAttachedToEngine(@NonNull FlutterPluginBinding binding) {
@@ -173,6 +175,9 @@ public class PurchasesFlutterPlugin implements FlutterPlugin, MethodCallHandler,
             case "getAppUserID":
                 getAppUserID(result);
                 break;
+            case "getStorefront":
+                getStorefront(result);
+                break;
             case "restorePurchases":
                 restorePurchases(result);
                 break;
@@ -275,6 +280,10 @@ public class PurchasesFlutterPlugin implements FlutterPlugin, MethodCallHandler,
             case "setFirebaseAppInstanceID":
                 String firebaseAppInstanceID = call.argument("firebaseAppInstanceID");
                 setFirebaseAppInstanceID(firebaseAppInstanceID, result);
+                break;
+            case "setTenjinAnalyticsInstallationID":
+                String tenjinAnalyticsInstallationID = call.argument("tenjinAnalyticsInstallationID");
+                setTenjinAnalyticsInstallationID(tenjinAnalyticsInstallationID, result);
                 break;
             case "setOnesignalID":
                 String onesignalID = call.argument("onesignalID");
@@ -483,6 +492,16 @@ public class PurchasesFlutterPlugin implements FlutterPlugin, MethodCallHandler,
         result.success(CommonKt.getAppUserID());
     }
 
+    private void getStorefront(final Result result) {
+        CommonKt.getStorefront(new Function1<Map<String, ? extends Object>, Unit>() {
+            @Override
+            public Unit invoke(Map<String, ?> storefrontMap) {
+                result.success(storefrontMap);
+                return null;
+            }
+        });
+    }
+
     private void restorePurchases(final Result result) {
         CommonKt.restorePurchases(getOnResult(result));
     }
@@ -609,6 +628,14 @@ public class PurchasesFlutterPlugin implements FlutterPlugin, MethodCallHandler,
 
     private void setFirebaseAppInstanceID(String firebaseAppInstanceId, final Result result) {
         SubscriberAttributesKt.setFirebaseAppInstanceID(firebaseAppInstanceId);
+        result.success(null);
+    }
+
+    private void setTenjinAnalyticsInstallationID(
+        String tenjinAnalyticsInstallationID,
+        final Result result
+    ) {
+        SubscriberAttributesKt.setTenjinAnalyticsInstallationID(tenjinAnalyticsInstallationID);
         result.success(null);
     }